extract block entity class
