+++ src/main/java/net/mrscauthd/boss_tools/gui/screens/blastfurnace/BlastFurnaceGui.java
@@ -13,8 +13,7 @@ import net.minecraftforge.items.SlotItemHandler;
 import net.minecraftforge.network.IContainerFactory;
 import net.mrscauthd.boss_tools.ModInnet;
 import net.mrscauthd.boss_tools.gui.helper.ContainerHelper;
-import net.mrscauthd.boss_tools.machines.BlastingFurnaceBlock;
-import net.mrscauthd.boss_tools.machines.BlastingFurnaceBlock.BlastingFurnaceBlockEntity;
+import net.mrscauthd.boss_tools.machines.tile.BlastingFurnaceBlockEntity;
 import net.mrscauthd.boss_tools.machines.tile.ItemStackToItemStackBlockEntity;
 
 public class BlastFurnaceGui {
@@ -36,7 +35,7 @@ public class BlastFurnaceGui {
 
 			IItemHandlerModifiable itemHandler = blockEntity.getItemHandler();
 			this.addSlot(new SlotItemHandler(itemHandler, ItemStackToItemStackBlockEntity.SLOT_INGREDIENT, 53, 19));
-			this.addSlot(new SlotItemHandler(itemHandler, BlastingFurnaceBlock.SLOT_FUEL, 53, 56));
+			this.addSlot(new SlotItemHandler(itemHandler, BlastingFurnaceBlockEntity.SLOT_FUEL, 53, 56));
 
 			this.addSlot(new SlotItemHandler(itemHandler, ItemStackToItemStackBlockEntity.SLOT_OUTPUT, 104, 38) {
 				@Override
